Add more transform tests
