--- test/test_transform/a--transform=test.yang
@@ -1,1 +1,0 @@
-a.yang
